packages/walkie-vision/pyproject.toml
[project]
name = "walkie-vision"
version = "0.1.0"
description = "Image Awareness Module (SAM 2.1 + YOLO11x + ArcFace)"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "torchvision",
    "ultralytics>=8.3.0",    # สำหรับ YOLO11
    "sam2",                  # สำหรับ SAM 2.1
    "insightface",           # สำหรับ ArcFace
    "onnxruntime-gpu",       # สำหรับรัน ArcFace/YOLO บน GPU
    "transformers",          # สำหรับ BLIP VQA
    "sentence-transformers",
    "numpy>=2.0.0",
    "opencv-python",
    "pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["walkie_vision"]
